+++ lib/rails_audit_log.rb
@@ -1,7 +1,5 @@
 require "rails_audit_log/version"
 require "rails_audit_log/engine"
-require "rails_audit_log/auditable"
-require "rails_audit_log/controller"
 
 module RailsAuditLog
   def self.actor